% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire distance and damage.}
\format{
A data frame with 15 rows and 2 variables
\describe{
  \item{DISTANCE}{measure of distance}
  \item{DAMAGES}{scale of damage}
  ...
}
}
\source{
canvas
}
\usage{
fire
}
\description{
15 lines of distance and damage
}
\keyword{datasets}
