% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddtfun.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{ddt data frame}

\item{SPECIES}{species of the fish}
}
\value{
plot and filtered data frame
}
\description{
myddt
}
\examples{
\dontrun{obj <- myddt(df = ddt, SPECIES = "CCATFISH")}
}
